chore: rename buckets file
